requirements.txt
asyncio==3.4.3
chaseinvest-api==0.2.4
discord.py==2.4.0
fennel-invest-api==1.1.0
firstrade==0.0.30
GitPython==3.1.43
public-invest-api==1.0.4
pyotp==2.9.0
python-dotenv==1.0.1
requests==2.32.3
# robin-stocks==3.1.0
-e git+https://github.com/NelsonDane/robin_stocks.git@2387113072bb26bcc1bd3fed8b4d74b0afb19f29#egg=robin-stocks
schwab-api==0.4.3
selenium-stealth==1.0.6
setuptools==74.0.0
tastytrade==8.2
vanguard-api==0.2.5
-e git+https://github.com/NelsonDane/webull.git@ef14ae63f9e1436fbea77fe864df54847cf2f730#egg=webull
streamlit==1.38.0
pandas==2.2.2
streamlit_extras==0.4.7
streamlit_authenticator==0.3.3
pyyaml==6.0.2

